Add option to preserve stateful data
Sometimes we want to redeploy the application configuration, such as the
earlier ImageMagick policy, without completely reinstalling the
operating system. Add an option to run through the deployment script
without touching the stateful data, which is the local MySQL database
and the files in the S3 bucket.

As part of this, move more of the configuration from files outside of
the repo into environment variables, and document using a template for
these environment variables.