Stop building web-app during vagrant provisioning
Building the web-app (formerly mdot) during `vagrant up` causes a lot of
problems. The most immediately obvious one is the time it takes;
building the web-app is a surprisingly compute- and I/O-heavy process,
and both of those resources are diminished inside a virtual machine
compared to the host operating system.

The bigger but more subtle problem is that running `vagrant up` breaks
the `node_modules` directory on the host: the guest is (almost
certainly) a different operating system, and `node_modules` is
non-portable because it contains platform-specific code. Worse, we're
running `npm install` with the argument `--no-bin-links`, which breaks
npm scripts.

After discussion, we decided that we're willing to accept the cost of a
slightly more difficult onboarding process for new developers in
exchange for a more productive development experience. This commit to
stop building mdot/web-app is the first implementation of that decision,
as web-app is under more-or-less constant development. We plan not to
add the notifications service to vagrant, and we may eventually move the
upload service out of vagrant if we find we need to do additional work
on it.

Remove the web-app build steps, document the additional manual steps
required, and reduce the memory allocation now that we're not building.

Issue #12 vagrant up breaks host node_modules