Mount back-end repo in Vagrant
When we consolidated api, daemon, library, and task-runner into a single
back-end repository, we continued to emulate the old layout by mounting
the subdirectories of back-end into their original locations in /data.
We did the same when packaging, so that we did not have to reconfigure
Apache.

In order to allow us to use other subdirectories in back-end, and to
support migrating from the disparate directories to the consolidated
directory in local/dev/staging/prod, mount back-end directly.

Once the AMIs and Vagrant are in alignment, we can update Apache to use
this new directory, but that work must be done later; this just sets the
stage.

More immediately, this allows developers to run PHPLint from within
Vagrant (instead of only on the host), and will enable us to add more
such tools.