Incorporate stela into the local environment
Before we start writing production code in stela, we need it to be part
of our local development environment. This commit adds a docker-compose
file which brings up stela, a postgres database to be shared by stela
and back-end, and an nginx server to properly route requests to either
back-end or stela.