Remove databases from Vagrant VM
Currently, our Vagrant VM runs a MySQL database and a PostgreSQL
database. We don't need the MySQL one anymore because we're not using
MySQL at all anymore, and we don't need the PostgreSQL one anymore
because local instances of our application now use a dockerized
PostgreSQL database. In light of this, this commit removes both
databases from the Vagrant VM.