Allow HTTP connections in the local environment
Presently, the nginx instance that routes requests to either back-end or
stela in the local environment only supports https. Because it would
take some effort to set up stela to be able to make https calls in the
local environment, and because this effort wouldn't have any meaningful
benefit, it is preferable to allow http calls. This commit updates the
docker compose and nginx.conf to do this.