Get whole database schema from back-end
Currently, the schema for the local database draws the starting-point
schema from `stela` and migrations from `back-end`. This was intended to
be part of incrementally moving things into `stela`, but it creates
needless complexity to make the database dependent on both those repos,
and it will be fairly easy to move both pieces to `stela` when `stela`
is ready to host migrations. In light of that, this commit changes the
local DB setup to use back-end for both the starting-point schema and
migrations.